api/requirements.txt
Django==4.1.1
psycopg2-binary==2.9.3
djangorestframework
drf-writable-nested
django-nested-inline